Converted everything over to use glm::vec3 and floats

+++ src/aabb.hpp
@@ -9,11 +9,11 @@ struct aabb {
   constexpr aabb(const point3 &min, const point3 &max) : min(min), max(max) {}
   constexpr aabb() : min(infinity), max(-infinity) {}
 
-  constexpr inline bool hit(const ray &r, double t_min, double t_max) const {
+  constexpr inline bool hit(const ray &r, float t_min, float t_max) const {
     for (int a = 0; a < 3; a++) {
-      const double invD = 1.0f / r.dir[a];
-      double t0 = (min[a] - r.orig[a]) * invD;
-      double t1 = (max[a] - r.orig[a]) * invD;
+      const float invD = 1.0f / r.dir[a];
+      float t0 = (min[a] - r.orig[a]) * invD;
+      float t1 = (max[a] - r.orig[a]) * invD;
       if (invD < 0.0f)
         std::swap(t0, t1);
       t_min = std::max(t0, t_min);
